Create README.md
# Steem reward manager

Srm is a tool to automatically redeem your steem rewards, and if they contain sbd, go to the internal market to buy steem with it. 

To use it, update the file "config.js" and put as many account as you want. Just don't forget that the script needs your active key.